refactor: centralize visibility controls
