+++ src/js/blocks/controls/buttons.js
@@ -1,33 +1,14 @@
 import { Fragment } from '@wordpress/element';
 import { InspectorControls } from '@wordpress/block-editor';
-import { PanelBody, ToggleControl } from '@wordpress/components';
+import { PanelBody } from '@wordpress/components';
+import { getVisibilityControls } from '../utils';
 
 export const controls = (BlockEdit, props) => (
     <Fragment>
         <BlockEdit {...props} />
         <InspectorControls>
             <PanelBody title="FlexLine Options">
-                <ToggleControl
-                    label="Hide on Desktop"
-                    checked={!!props.attributes.hideOnDesktop}
-                    onChange={(newValue) =>
-                        props.setAttributes({ hideOnDesktop: newValue })
-                    }
-                />
-                <ToggleControl
-                    label="Hide on Tablet"
-                    checked={!!props.attributes.hideOnTablet}
-                    onChange={(newValue) =>
-                        props.setAttributes({ hideOnTablet: newValue })
-                    }
-                />
-                <ToggleControl
-                    label="Hide on Mobile"
-                    checked={!!props.attributes.hideOnMobile}
-                    onChange={(newValue) =>
-                        props.setAttributes({ hideOnMobile: newValue })
-                    }
-                />
+                {getVisibilityControls(props)}
             </PanelBody>
         </InspectorControls>
     </Fragment>
