+++ src/js/blocks/controls/visibility.js
@@ -1,35 +1,16 @@
 import { Fragment } from '@wordpress/element';
 import { InspectorControls } from '@wordpress/block-editor';
-import { PanelBody, ToggleControl } from '@wordpress/components';
+import { PanelBody } from '@wordpress/components';
+import { getVisibilityControls } from '../utils';
 
 export const controls = (BlockEdit, props) => (
     <Fragment>
-        <BlockEdit {...props} />
-        <InspectorControls>
-            <PanelBody title="FlexLine Visibility">
-                <ToggleControl
-                    label="Hide on Desktop"
-                    checked={!!props.attributes.hideOnDesktop}
-                    onChange={(newValue) =>
-                        props.setAttributes({ hideOnDesktop: newValue })
-                    }
-                />
-                <ToggleControl
-                    label="Hide on Tablet"
-                    checked={!!props.attributes.hideOnTablet}
-                    onChange={(newValue) =>
-                        props.setAttributes({ hideOnTablet: newValue })
-                    }
-                />
-                <ToggleControl
-                    label="Hide on Mobile"
-                    checked={!!props.attributes.hideOnMobile}
-                    onChange={(newValue) =>
-                        props.setAttributes({ hideOnMobile: newValue })
-                    }
-                />
-            </PanelBody>
-        </InspectorControls>
+	<BlockEdit {...props} />
+	<InspectorControls>
+	    <PanelBody title="FlexLine Visibility">
+		{getVisibilityControls(props)}
+	    </PanelBody>
+	</InspectorControls>
     </Fragment>
 );
